requirements.txt
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
python-dotenv==1.0.0